Bug fix & Beamcoder Removed
Normalized all variable references to script names, so that they start with a lowercase letter.

OnUserDefined --> onUserDefined.

This fixes Module and Area userDefined scripts not firing because the engine was trying to target onUserDefined when it was wrongly named OnUserDefined instead.

Removed beamcoder from the project. I hope to write a BIK decoder to replace it in the future.