Engine Object Visibility Optimization
The engine is now using a raycaster on each object to determine which room it resides in.

The visibility of the object is dependant on the visibility of it's current room.

If the object is not visible it can still run scripts and move around. It just wont be rendered to the screen, which increases performance.

The room visibility manager is now more accurate.