Added the ModulePlayer class
Some things like hitpoint calculations seem to be handled differently between Creatures and the Player.

For instance the players currentHitPoints seem to track damage as a positive value past 0 and temporary hitpoints healed above the max as a negative value below 0. So 0 currentHitPoints means the player is at full health.

Creatures and PartyMembers seem to track currentHitPoints as the actual number of remaining hitpoints so 0 currentHitPoints actually means the creature is dead.