Update: OdysseyController
Added a new class called OdysseyController

All model controllers are now based on this class and are located in the js/aurora/controllers directory.

I will eventually be renaming all Aurora references to Odyssey. The aurora references are a holdout from the early days of KotOR modding when the only documentation available was for the Aurora Engine. Because of that the KotOR Modding Community reused the same names as the NWN Modding Community.