update: GameEffects
Added code so that GameEffects cannot be applied more than once.

Added support for EffectIcon.

EffectLinks now unwrap their effects onto the object's effects array when using ModuleObject.AddEffect. This aligns more with what i'm seeing in the EffectList data in savegames.

So if you try to add an EffectLink to an object. AddEffect will push the effects inside of the EffectLink instead. Since EffectLinks can also have EffectLinks inside it will keep recursing until all effects are added.

NWScript: When SetEffectIcon is called it now creates an EffectIcon and uses a EffectLink  to link the supplied and the new effect icon together. the EffectLink is returned.