Update: Combat Melee on Melee Animations
If the creature is using a basic melee attack, and the target creature is using a melee weapon, and they are both directly attacking each other. Then they use variations of c*a*, c*d*, and c*p*.

This took me a while to figure out because I thought they were controlled by the combatanimations.2da. It appears that all other attacks are handled except for this particular senario. I believe this was hardcoded to act this way in the original game, and this is why they are not present in the combatanimations.2da