Update ModuleCreatureController.js
When damage is recieved it now goes into the object's damageList. If there is delay applied it will only apply once the delay has expired. I'm doing this to try and sync up animations and damage.

Modified the ANIMATE action to support the changes to actionPlayAnimation().

Added the isDueling() method to detect when creatures are in melee v melee combat.

Added a check of isSimpleCreature() to getCombatAnimationWeaponType(), because simple creatures without claw items were not being handled right.

Fixed the flourish() method