Update CombatEngine.js
Added support for EffectAssuredHit.

Added support for attack damage delay to sync it with the attack animation.

Modified the dueling timer to better keep the combatant and it's targets timers in sync.

Modified the attack animation to use overlayAnimation instead of applying it directly to the model.

Moved the pain sounds to the damageList update area of ModuleCreatureController.