Fix: INI file missing on some installations
It appears some distributions of KotOR and TSL do not have INI files. This was causing fatal errors in launching KotOR.js. I rewrote the INIConfig.js file to take in a JSON object to populate it with default values. If an INI file is present the values will overwrite the default ones.

Thank you Canderis for bringing this to my attention!