REFEACTOR: AuroraModel
The code for building AuroraModelNodes has now be moved outside of AuroraModel and into the AuroraModelNode classes for each node type. I originally intended this to work this way, but i've just been too lazy to rip it all apart.

So mesh building is now in AuroraModelNodeMesh and so on.

AuroraModelNodes take advantage of inheritance. So AuroraModelNodeSkin doesn't include any of the Mesh code, becuase that code is in AuroraModelNodeMesh. This works because AuroraModelNodeSkin inherits the AuroraModelNodeMesh class, and AuroraModelNodeMesh inherits the AuroraModelNode class.

So when the readBinary() method is called on AuroraModelNodeSkin it goes: AuroraModelNode.readBinary() -> AuroraModelNodeMesh.readBinary() -> AuroraModelNodeSkin.readBinary().