BugFix: EffectDisguise Stays After Unequip Disguise
Reworked the GameEffect initialization to separate setCreator() and the new setAttachedObject() methods. Before setCreator() was also setting the this.object property which wasn't correct.

this.creator is a reference to the object that created the effect
this.object is a reference to the object the effect is attached to

Made some changes to how EffectDisguise is initialized and how it is checked when a creature object calls getAppearance()