Add: Screen Size Aware Black Void Menu Background
Added the backgroundVoidSprite to the game Menu class. When a menu is created and sets voidFill to true, the Menu class will create a black sprite that sits behind the Menu. The sprite with always be the width and height of the game window. This keeps certain in game menus covering the entire screen so that the in game world isn't visible on resolutions larger than 1600 x 1200