Initial Controller Support
After connecting a GamePad press one button on the controller to update the GamePad handler. You can then choose your controller from the GamePads menu at the top of the window.

The only supported controls at this time are analog movement and camera movement with the right analog stick.

Action Classes are loaded but not in use.