add: Video Support
Added video support by converting all .bik videos to .mp4 on launch and using the HTMLVideoElement for playback on screen.

Long term goal would be to replace this with a full .bik decoder and playback solution.

When you try to launch a game from the launcher, it will now check the movies folder to see if there are any .bik files that have not been converted to .mp4. If it finds any it will convert them and then launch the game.