fix: collision warping inside placeables and off of walkable faces
creatures now only check against placeables and doors that are inside the same room.

after a collision has been detected check to see if the resulting position will land the creature inside of a placeable. If so cancel the movement for this frame.

If a creatures position is not on a valid walkable space after collision then cancel the movement for this frame.